Tweaking token behavior

+++ requirements/mura/client/api/oath/oauthClientBean.cfc
@@ -22,21 +22,33 @@ component extends="mura.bean.beanORM" entityName='oauthClient' table="toauthclie
     function generateToken(granttype='client_credentials'){
         var token=getBean('oauthToken');
 
-        token.set({
-            clientid=get('clientid'),
-            granttype=arguments.granttype
-        }).save();
+        var existingTokens=token.getFeed()
+            .where('clientid').isEQ(get('clientid'))
+            .andProp('granttype').isEQ(arguments.granttype)
+            .andProp('expires').isGT(now())
+            .getIterator();
 
-        if(arguments.granttype =='client_credentials'){
-            var expiredTokens=token.getFeed()
-                .where('clientid').isEQ(get('clientid'))
-                .andProp('granttype').isEQ('client_credentials')
-                .andProp('expires').isLT(dateAdd('d',-1,now()))
-                .getIterator();
+        if(existingTokens.hasNext()){
+            return existingTokens.next();
+        } else {
 
-            if(expiredTokens.hasNext()){
-                while(expiredTokens.hasNext()){
-                    expiredTokens.next().delete();
+
+            token.set({
+                clientid=get('clientid'),
+                granttype=arguments.granttype
+            }).save();
+
+            if(arguments.granttype =='client_credentials'){
+                var expiredTokens=token.getFeed()
+                    .where('clientid').isEQ(get('clientid'))
+                    .andProp('granttype').isEQ('client_credentials')
+                    .andProp('expires').isLT(dateAdd('d',-1,now()))
+                    .getIterator();
+
+                if(expiredTokens.hasNext()){
+                    while(expiredTokens.hasNext()){
+                        expiredTokens.next().delete();
+                    }
                 }
             }
         }
